+++ components/Collapse/Collapse.js
@@ -2,13 +2,14 @@
  * Created by kisnows on 2017/2/17.
  */
 import React, {Component, PropTypes} from 'react'
-import List from './List'
 
-export default class CollapseList extends React.Component {
+export default class Collapse extends React.Component {
   constructor(props) {
     super(props)
     this.state = {
+      // 当前打开元素的集合
       openListIdCollection: this.props.openListIdCollection || [],
+      // SubComponent 渲染所需要的数据集合
       listCollection: this.props.listCollection || []
     }
   }
@@ -16,23 +17,22 @@ export default class CollapseList extends React.Component {
   static propTypes = {
     prefixCls: PropTypes.string,
     openListIdCollection: PropTypes.arrayOf(PropTypes.number),
-    listCollection: PropTypes.arrayOf(PropTypes.shape({
-      title: PropTypes.node,
-      content: PropTypes.node,
-      isOpen: PropTypes.bool
-    })).isRequired,
+    listCollection: PropTypes.array.isRequired,
+    // 手风琴模式
     accordion: PropTypes.bool,
-    onListChange: PropTypes.func
+    onListChange: PropTypes.func,
+    subComponent: PropTypes.element
   }
 
   static defaultProps = {
     prefixCls: 'NEUI',
   }
+
   handleListChange = (id, collapse) => {
     //TODO 把 openListIdCollection 改为不可重复的数据结构 set
     let {openListIdCollection, listCollection}= this.state
     const {accordion} = this.props
-    listCollection[id]['collapse'] = collapse
+    listCollection[id]['isCollapse'] = collapse
 
     function removeCollapseListId() {
       return openListIdCollection = openListIdCollection.filter((value) => {
@@ -59,20 +59,21 @@ export default class CollapseList extends React.Component {
   }
   renderList = () => {
     const {openListIdCollection, listCollection} = this.state
+    const subComponent = this.props.subComponent
     return listCollection.map((v, k) => {
-      const isCollapse = openListIdCollection.indexOf(k) === -1
-      return <List collapse={isCollapse}
-                   title={v.title}
-                   content={v.content}
-                   onChange={this.handleListChange}
-                   id={k}
-                   key={k}/>
+      v.isCollapse = openListIdCollection.indexOf(k) === -1
+      return React.cloneElement(subComponent, {
+        data: v,
+        onChange: this.handleListChange,
+        id: k,
+        key: k,
+      })
     })
   }
 
   render() {
     const Lists = this.renderList()
-    const {prefixCls, openListIdCollection, listCollection, accordion, onListChange, ...others} = this.props
+    const {prefixCls, openListIdCollection, listCollection, accordion, onListChange, subComponent, ...others} = this.props
     return (
       <div {...others}>
         {Lists}
