+++ components/style/_base.scss
@@ -8,7 +8,7 @@
 }
 
 html {
-  background-color: $gray;
+  background-color: #f8f8f8;
   font-size: pxToRem(28)
 }
