+++ examples/demo/Alert/index.js
