+++ examples/demo/App.jsx
@@ -1,12 +1,11 @@
-import React, { PropTypes } from 'react'
-import { Link } from 'react-router'
+import React, {PropTypes} from 'react'
+import {Link} from 'react-router'
 
 class App extends React.Component {
 
   render() {
     return (
       <div>
-        <Link to="/">主页</Link>
         {this.props.children}
       </div>
     )
