+++ examples/demo/Form/index.js
@@ -2,11 +2,31 @@
  * Created by kisnows on 2016/12/26.
  */
 import React from 'react'
-import TestForm from '../../../components/Form/Form'
-import Input from '../../../components/Form/Input'
-import {Button, FormCell, CellHeader, Toast} from '../../../components'
+import {
+  Form,
+  FormCell,
+  Input,
+  Select,
+  CheckBox,
+  Cells,
+  Cell,
+  CellHeader,
+  Icon,
+  CellBody,
+  CellFooter,
+  Panel,
+  DatePicker,
+  Picker,
+  Modal,
+  Alert,
+  VerifyButton,
+  Toast,
+  Button,
+  Collapse
+} from '../../../components'
 import Logger from '../../../utils/log'
-
+import lang from '../../utils/lang'
+import validate from '../../utils/validate'
 
 const logger = new Logger('DEBUG', 'FormDemo')
 
@@ -89,16 +109,27 @@ class FormDemo extends React.Component {
   render() {
     const {showInput, msg, showToast, formData} = this.state
     const {isComplete} = formData
+    const genderData = [{
+      name: '男',
+      value: 0,
+      disabled: true
+    }, {
+      name: '女',
+      value: 1
+    }]
     logger.info('render', this.state.formData)
     console.log(isComplete)
     return (
       <section className="page-form-demo">
+        <div className="page--header">
+          <h1 className="page-title">Form</h1>
+          <p className="page--desc">表单组件，包括 Form, Input, Select, Checkbox。<br/> Form 组件包含了填写校验，错误提示等功能。</p>
+        </div>
         <Toast content={msg}
                show={showToast}
                onClose={this.closeToast}
         />
-        <p>page-form-demo</p>
-        <TestForm
+        <Form
           onSubmit={this.handleSubmit}
           onFieldChange={this.handleFieldChange}
           onChange={this.handleChange}
@@ -108,19 +139,42 @@ class FormDemo extends React.Component {
         >
           <FormCell>
             <CellHeader>Name</CellHeader>
-            <_FieldInput type='text'
-                         name='Name'
+            <Input type='text'
+                   name='name'
+                   errorMsg={lang.nameErrorMsg}
+                   validate={validate.name}
             />
           </FormCell>
-          <FormCell cellsEnd>
+          <FormCell>
             <CellHeader>Phone</CellHeader>
-            <_FieldInput type='text'
-                         name='phone'
-                         validate={/^\d{11}$/}
+            <Input type='tel'
+                   name='phone'
+                   validate={validate.phone}
+                   errorMsg={lang.phoneErrorMsg}
             />
+          </FormCell>
+          <FormCell>
+            <CellHeader>Verify Code</CellHeader>
+            <Input type='number'
+                   name='verifyCode'
+                   errorMsg={lang.smsCodeErrorMsg}
+                   validate={/\d{4}/}
+            />
+            <CellFooter><VerifyButton/></CellFooter>
+          </FormCell>
+          <FormCell>
+            <CellHeader>Gender</CellHeader>
+            <Select name='gender'
+                    data={genderData}
+                    required={false}
+            />
+          </FormCell>
+          <FormCell>
+            <CellBody><label htmlFor="is">Is yourself?</label></CellBody>
+            <CellFooter><CheckBox name='is' id="is"/></CellFooter>
           </FormCell>
           <Button type="submit" disabled={!isComplete}>提交</Button>
-        </TestForm>
+        </Form>
       </section>
     )
   }
