+++ examples/demo/Home/Home.scss
@@ -1,0 +1,5 @@
+@import "../../style/config";
+
+.page-home {
+
+}
