+++ examples/demo/Home/index.js
@@ -1,7 +1,7 @@
 /**
  * Created by kisnows on 2016/9/21.
  */
-import HomeDemo from './HomeDemo'
-import './HomeDemo.scss'
+import HomeDemo from './Home'
+import './Home.scss'
 
 export default HomeDemo
