--- examples/demo/HomeDemo/HomeDemo.scss
@@ -1,84 +1,0 @@
-/*@import '../../scss/setting';
-
-.page_home {
-  header {
-    background-color: $white;
-    margin-bottom: pxToRem(24);
-    border-bottom: 1px solid $border_primary;
-  }
-  .before_loan {
-    margin-bottom: pxToRem(24);
-  }
-  .after_loan {
-    height: pxToRem(290);
-    text-align: center;
-    padding-top: pxToRem(85);
-    .repay_date {
-      color: $text_highlight;
-      font-size: pxToRem(24);
-    }
-    .need_repaid {
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      margin-top: pxToRem(20);
-      font-size: pxToRem(80);
-      color: $text_highlight;
-    }
-    .not_repaid {
-      margin-top: pxToRem(72);
-      color: $text_secondary;
-      font-size: pxToRem(24);
-      strong {
-        color: $text_highlight;
-      }
-    }
-  }
-  .order_state {
-    height: pxToRem(260);
-    background-color: $white;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    color: $blue;
-    font-size: pxToRem(44);
-  }
-  .detail {
-    display: none;
-    height: pxToRem(110);
-    background-color: $white;
-    padding: pxToRem(25) 0;
-    border-top: 1px solid $border_primary;
-    &.active {
-      display: flex;
-    }
-    li {
-      flex: 1;
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-      justify-content: space-between;
-      position: relative;
-      span {
-        color: $text_primary;
-        &:first-child {
-          color: $text-secondary;
-        }
-      }
-      &:after {
-        content: '';
-        position: absolute;
-        right: 0;
-        top: 50%;
-        transform: translate3d(0, -50%, 0);
-        display: block;
-        height: 60px;
-        width: 1px;
-        background-color: #eaeaea;
-      }
-    }
-    li:last-child:after {
-      display: none;
-    }
-  }
-}*/
