+++ examples/style/_global.scss
@@ -1,0 +1,13 @@
+a {
+  text-decoration: none;
+  color: $blue;
+}
+
+.page--header {
+  padding: pxToRem(30);
+}
+
+.page--desc {
+  font-size: $normalSize;
+  line-height: 1.6;
+}
