+++ examples/style/_utils.scss
@@ -1,0 +1,23 @@
+@function pxToRem($px) {
+  @return ($px/40)*1rem
+}
+
+@mixin flexCenter {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+@mixin iconfont-mixin() {
+  display: inline-block;
+  font-style: normal;
+  font-size: pxToRem(22);
+  -webkit-font-smoothing: antialiased;
+  -webkit-text-stroke-width: 0.2px;
+  -moz-osx-font-smoothing: grayscale;
+  margin-left: pxToRem(10);
+  &:before {
+    display: block;
+    font-family: "iconfont" !important;
+  }
+}
