+++ examples/style/app.scss
@@ -1,1 +1,2 @@
-@import "setting";
+@import "config";
+@import "global";
