--- examples/style/setting.scss
@@ -1,7 +1,0 @@
-@import "../components/style/config";
-@import "../components/style/utils";
-//-----------------global setting---------------
-html {
-  background-color: $gray;
-  font-size: pxToRem(28)
-}
